[SB-78] feat:TopicIssue 테스트 코드 추가

+++ backend/src/main/java/com/sisibibi/api/domain/roomparticipant/dto/response/RoomParticipantCountRes.java
@@ -2,6 +2,6 @@ package com.sisibibi.api.domain.roomparticipant.dto.response;
 
 public record RoomParticipantCountRes(
     Long roomId,
-    long participantCount
+    int participantCount
 ) {
 }
